% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{fetchAndWrangleIU}
\alias{fetchAndWrangleIU}
\title{Fetch IU data from AGOL and do preliminary data wrangling}
\usage{
fetchAndWrangleIU(
  iu_urls =
    c("https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_Database/FeatureServer",
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_Site/FeatureServer",
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_UnknownPlant/FeatureServer",
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_RepeatPhotos/FeatureServer"),
  agol_username = "mojn_data"
)
}
\arguments{
\item{agol_username}{Authentication token (not needed for public layers)}

\item{urls}{list of IU database URLs}
}
\value{
A list of data frames and metadata
}
\description{
Fetch IU data from AGOL and do preliminary data wrangling
}
