% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchFeatureService}
\alias{fetchFeatureService}
\title{Fetch feature service info from AGOL}
\usage{
fetchFeatureService(url, token)
}
\arguments{
\item{url}{Feature service URL}

\item{token}{Authentication token (not needed for public layers)}
}
\value{
A list
}
\description{
Retrieves metadata from AGOL layers and tables.
}
