% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadAndWrangleIU}
\alias{loadAndWrangleIU}
\title{Fetch IU data from AGOL and do preliminary data wrangling}
\usage{
loadAndWrangleIU(
  iu_urls =
    c("https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_Database/FeatureServer",
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_Site/FeatureServer",
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_UnknownPlant/FeatureServer",
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_IU_RepeatPhotos/FeatureServer"),
  agol_username = "mojn_data",
  tables_to_remove = c("Site - DEPRACATED"),
  cols_to_remove = c("objectid", "InstanceName", "^app_.*", "GapsKey", "^Shrub.*")
)
}
\arguments{
\item{agol_username}{Authentication token (not needed for public layers)}

\item{IU_urls}{list of IU database URLs}
}
\value{
A list of data frames and metadata
}
\description{
Fetch IU data from AGOL and do preliminary data wrangling
}
